+++ src/main/kotlin/cloud/glitchdev/rfu/gui/window/PartyFinderWindow.kt
@@ -20,6 +20,7 @@ import cloud.glitchdev.rfu.gui.components.elementa.group.GroupManager
 import cloud.glitchdev.rfu.gui.components.partyfinder.UICreateParty
 import cloud.glitchdev.rfu.gui.components.partyfinder.UIFilterArea
 import cloud.glitchdev.rfu.gui.components.partyfinder.UIPartyCard
+import cloud.glitchdev.rfu.gui.components.partyfinder.UIPartyPresetsModal
 import cloud.glitchdev.rfu.model.party.FishingParty
 import cloud.glitchdev.rfu.utils.Coroutines
 import cloud.glitchdev.rfu.utils.World
@@ -32,6 +33,7 @@ import gg.essential.elementa.components.UIContainer
 import gg.essential.elementa.components.UIImage
 import gg.essential.elementa.components.UIRoundedRectangle
 import gg.essential.elementa.components.UIText
+import gg.essential.elementa.components.inspector.Inspector
 import gg.essential.elementa.constraints.AspectConstraint
 import gg.essential.elementa.constraints.CenterConstraint
 import gg.essential.elementa.constraints.ChildBasedSizeConstraint
@@ -68,6 +70,7 @@ object PartyFinderWindow : BaseWindow(false), Feature {
     private var isPeeking = false
 
     lateinit var popup: UIPopup
+    lateinit var presetsModal: UIPartyPresetsModal
     lateinit var filterButton : UIButton
     lateinit var refreshButton : UIButton
     lateinit var filterArea : UIContainer
@@ -161,6 +164,9 @@ object PartyFinderWindow : BaseWindow(false), Feature {
         createPartyCreationArea(contentWrapper)
         createPartyArea(contentWrapper)
 
+        presetsModal = UIPartyPresetsModal(5f).childOf(window)
+        presetsModal.hideModal()
+
         popup = UIPopup(5f, "", isBordered = true).childOf(window).colors {
             primaryColor = UIScheme.pfCardBorder.toConstraint()
             innerColor = UIScheme.pfCardBg.toConstraint()
@@ -199,6 +205,9 @@ object PartyFinderWindow : BaseWindow(false), Feature {
 
         val createImage = UIImage.ofResource("/assets/rfu/ui/edit.png")
         UIButton.withImage(createImage, 5f) {
+            if (creationOpen && ::creationArea.isInitialized) {
+                creationArea.saveSessionState()
+            }
             creationOpen = !creationOpen
             onUpdate()
         }.constrain {
@@ -390,5 +399,11 @@ object PartyFinderWindow : BaseWindow(false), Feature {
             }
         }
         updateFiltering()
+    }
+
+    override fun onWindowClose() {
+        if (::creationArea.isInitialized) {
+            creationArea.saveSessionState()
+        }
     }
 }